Se arreglaron los resultados del JSON y cambiamos unas clases para que coincidieran con los valores que se calcula

+++ MovilidadMedellin/src/resultadosSimulacion/Distancias.scala
@@ -1,5 +1,5 @@
 package resultadosSimulacion
 
-case class Distancias(minima: Int, maxima: Int, promedio: Int) {
+case class Distancias(minima: Int, maxima: Int, promedio: Double) {
   
 }
