+++ MovilidadMedellin/src/resultadosSimulacion/VehiculosInterseccion.scala
@@ -1,5 +1,5 @@
 package resultadosSimulacion
 
-case class VehiculosInterseccion(promedioOrigen: Int, promedioDestino: Int, sinOrigen: Int, sinDestino: Int) {
+case class VehiculosInterseccion(promedioOrigen: Double, promedioDestino: Double, sinOrigen: Int, sinDestino: Int) {
   
 }
